+++ packages/app/src/systems/Ecosystem/components/ProjectList/ProjectList.tsx
@@ -1,12 +1,7 @@
 import { cssObj } from '@fuel-ui/css';
 import { Grid } from '@fuel-ui/react';
-import { AnimatePresence, motion } from 'framer-motion';
 
 import { ProjectItem, type Project } from '../ProjectItem';
-
-import { animations } from '~/systems/Core';
-
-const MotionGrid = motion(Grid);
 
 type ProjectListProps = {
   projects: Project[];
@@ -14,21 +9,16 @@ type ProjectListProps = {
 
 export const ProjectList = ({ projects }: ProjectListProps) => {
   return (
-    <MotionGrid
+    <Grid
       gap="$8"
       templateColumns="repeat(2, 1fr)"
       templateRows="repeat(2, 1fr)"
       css={styles.grid}
-      {...animations.slideInBottom({
-        transition: { staggerChildren: 0.5, type: 'spring' },
-      })}
     >
-      <AnimatePresence initial={false} mode="sync">
-        {projects?.map((project) => (
-          <ProjectItem {...project} key={project.url} />
-        ))}
-      </AnimatePresence>
-    </MotionGrid>
+      {projects.map((project) => (
+        <ProjectItem {...project} key={project.url} />
+      ))}
+    </Grid>
   );
 };
