+++ packages/app/src/systems/Ecosystem/components/ProjectList/index.tsx
@@ -1,0 +1,1 @@
+export * from './ProjectList';
