+++ packages/app/src/systems/Ecosystem/components/index.tsx
@@ -1,0 +1,3 @@
+export * from './ProjectItem';
+export * from './ProjectList';
+export * from './EcosystemTags';
