+++ packages/app/src/systems/Ecosystem/hooks/index.tsx
@@ -1,0 +1,1 @@
+export * from './useEcosystem';
